Reorganize MVPA scripts
